feat: Refactor Protected component and add AdminProtected for role-based access control
