+++ client/app/admin/page.tsx
@@ -1,0 +1,7 @@
+
+
+export default function AdminHome() {
+    return (
+        <div></div>
+    )
+}
