+++ client/components/utils/protected/Protected.tsx
@@ -1,9 +1,9 @@
 "use client";
 
 import { useEffect } from "react";
-import { useContextFunc } from "./context/AppContext";
+import { useContextFunc } from "../../context/AppContext";
 import { useRouter } from "next/navigation";
-import Loader from "./Loader/Loader";
+import Loader from "../../Loader/Loader";
 
 export default function Protected({ children }: { children: React.ReactNode }) {
   const { loadingContext, accessToken } = useContextFunc();
